Revert "Implement security for IIIF images (#415)"
This reverts commit 9a631802b31d01f332792f49a28d1853d2354763.
