+++ app/controllers/manifests_controller.rb
@@ -2,8 +2,8 @@
 
 # Takes a request for a manifest/oid and stream the JSON for that oid from S3
 class ManifestsController < ApplicationController
+  before_action :check_authorization
   include Blacklight::Catalog
-  include CheckAuthorization
 
   def show
     remote_path = pairtree_path
@@ -24,5 +24,25 @@ class ManifestsController < ApplicationController
     client = Aws::S3::Client.new
     response = client.get_object(bucket: ENV['SAMPLE_BUCKET'], key: remote_path)
     response.body&.read
+  end
+
+  def check_authorization
+    @response, @document = search_service.fetch(params[:id])
+
+    # Handle when the 'visibility_ssi' key doesn't exist on the manifest
+    unless @document.key?('visibility_ssi')
+      render json: { error: 'not-found' }.to_json, status: 404
+      return false
+    end
+
+    case @document['visibility_ssi']
+    when 'Public'
+      true
+    when 'Yale Community Only'
+      return true if current_user
+
+      render json: { error: 'not-found' }.to_json, status: 404
+      false
+    end
   end
 end
